requirements.txt
Django==5.0.3
djangorestframework-simplejwt==5.3.1